Envs

PathPlanning.Envs.add_obstacle! — Method

add_obstacle!(env::Env, obstacle::AbstractObstacle)

Add a obstacle to Env object.

PathPlanning.Envs.create_example_2D_env — Method

create_example_2D_env()::Env

Return an example Env object that includes 4 RectObstacles and 5 CircleObstacles.

PathPlanning.Envs.is_inside — Method

is_inside(obs::CircleObstacle{N},, position::SVector{N,Float64})::Bool where {N}

Return if the position is inside the CircleObstacle.

PathPlanning.Envs.is_inside — Method

is_inside(obs::RectObstacle{N}, position::SVector{N,Float64})::Bool where {N}

Return if the position is inside the RectObstacle.

PathPlanning.Envs.is_inside_any_obstacle — Method

is_inside_any_obstacle(env::Env, position::SVector{N,Float64})::Bool where {N}

Return if the position is inside any obstacle in the Env.

PathPlanning.Envs.CircleObstacle — Type

CircleObstacle{N}

A Circle-type Obstacle.

Fields

center::SVector{N,Float64}: center position of the circle-type obstacle.

radius::Float64: radius of the circle-type obstacle.

PathPlanning.Envs.Env — Type

Env

A Env object has all obstacles that are in the environment.

Fields

num_obstacles::Int64: the number of the obstacles

obstacles::Vector{AbstractObstacle}: obstacles that have a position and a size

PathPlanning.Envs.Env — Method

Env(obstacles::Vector{AbstractObstacle})

Constructor for Env with obstacles.

PathPlanning.Envs.Env — Method

Env()

Constructor for Env with no obstacles.

PathPlanning.Envs.RectObstacle — Type

RectObstacle{N}

A Rectangle-type Obstacle.

Fields

center::SVector{N,Float64}: center position of the rectangle-type obstacle.

size::SVector{N,Float64}: width of each side of the rectangle-type obstacle.