Planners

PathPlanning.Planners.add_node! — Method

add_node!(rrt_star::RRTStar{N}, new_node::Node{N}, parent_node_index::Int64)::Nothing where {N}

Add new node with the parent index and the cost to the graph. The cost of the new node is the cost of the parent node and the distance between the parent node and the new node.

PathPlanning.Planners.calc_distance — Method

calc_distance(node1::Node{N}, node2::Node{N})::Float64 where {N}

Calculate the Euclidean distance between the two nodes in the N-dimensional space.

PathPlanning.Planners.extract_path — Method

extract_path(rrt::AbstractRRT{N}) where N

Return a path from start node to goal node by extracting nodes.

Arguments

rrt::AbstractRRT{N}: a planner using RRT based algorithm

PathPlanning.Planners.get_extended_node — Method

get_extended_node(rrt::RRT{N}, nearest_node::Node{N}, new_node::Node{N})::Node{N} where {N}

Create and return a node at the position from the nearest node to the new node as close as by the step size.

Arguments

rrt::AbstractRRT{N}: a planner using RRT based algorithm

nearest_node::Node{N}: the node nearest to the new node in the graph

new_node::Node{N}: the node newly sampled

PathPlanning.Planners.get_near_node_indices — Method

get_near_node_indices(rrt::RRTStar{N}, new_node::Node{N}, distance::Float64)::Vector{Int64} where {N}

Get nodes within distance from the new_node.

PathPlanning.Planners.get_nearest_node_index — Method

get_nearest_node_index(rrt::AbstractRRT{N}, new_node::Node{N})::Int64 where {N}

Return the index of the node nearest to the given node in the graph.

PathPlanning.Planners.get_shortest_parent_node_index — Method

get_shortest_parent_node_index(rrt::RRTStar{N}, new_node::Node{N})::Int64 where {N}

Get the parent node that minimizes the length between the start node and the new node.

PathPlanning.Planners.is_near_the_goal — Method

is_near_the_goal(rrt::AbstractRRT{N}, node::Node{N}) where {N}

Return if given node is near the goal (distance from goal is smaller than the rrt.step_size.)

Arguments

rrt::AbstractRRT{N}: a planner using RRT based algorithm

node::Node{N}: a node in the search space

PathPlanning.Planners.plan — Method

plan(rrt_star::RRTStar{N})::Vector{Node{N}} where {N}

Find and return a path from the start node rrt_star.start to the goal node rrt_star.goal with RRT* algorithm.

Arguments

rrt_star::RRTStar{N}: a planner using RRT*

Returns

path::Vector{Node{N}}: path (sequence of nodes) from the start node rrt.start to the goal node rrt.goal

PathPlanning.Planners.plan — Method

plan(rrt::RRT{N})::Vector{Node{N}} where {N}

Find and return a path from the start node rrt.start to the goal node rrt.goal with RRT algorithm.

Arguments

rrt::RRT{N}: a planner using RRT

Returns

path::Vector{Node{N}}: path (sequence of nodes) from the start node rrt.start to the goal node rrt.goal

PathPlanning.Planners.rewire_near_nodes! — Method

rewire_near_nodes!(rrt_star::RRTStar{N}, near_node_indices::Vector{Int64}, new_node_index::Int64) where {N}

Change parent nodes to the new node if the new cost is smaller than the current cost.

PathPlanning.Planners.sample — Method

sample(rrt::AbstractRRT{N})::Node{N} where {N}

Sample a node from the search space.

PathPlanning.Planners.Node — Type

Node{N}

Node configures a graph.

Fields

position::SVector{N, Float64}: position of the node in the search space

parent::Union{Int64, Nothing}: index of the parent node

PathPlanning.Planners.Node — Method

Node(position::SVector{N, Float64}) where {N}

Constructor of the Node struct. The index of the parent node is initiated as nothing.

Arguments

position::SVector{N, Float64}: position of the node in the search space

PathPlanning.Planners.RRT — Type

RRT{N}

A planner using RRT.

Fields

start::Node{N}: start position in the search space

goal::Node{N}: goal position in the search space

low::SVector{N,Float64}: lower bounds of the search space

high::SVector{N,Float64}: upper bounds of the search space

nodes::Vector{Node{N}}: nodes configure tje graph

goal_sample_rate::Float64: rate of sampling the goal node

step_size::Float64: maximum distance between each node

max_iter::Int64: maximum the number of the iterations

is_approved::Union{Function, Nothing}: a function that returns if the node can be added the graph

enable_logging::Bool: whether to log (is_goaled, RRTStar instance) for each steps in planning

logs::Vector{Dict{String, Any}}: (is_goaled, RRT instance) for each steps in planning

PathPlanning.Planners.RRT — Method

RRT(
    start::SVector{N,Float64},
    goal::SVector{N,Float64},
    low::SVector{N,Float64},
    high::SVector{N,Float64};
    goal_sample_rate::Float64 = 0.2,
    step_size::Union{Float64,Nothing} = nothing,
    max_iter::Int64 = 500,
    is_approved::Union{Function, Nothing} = nothing,
    enable_logging::Bool = false
) where {N}

Constructor of RRT struct. The step_size is initiated as 1/20th of the distance between the start node and the start node by default.

Arguments

start::Node{N}: start position in the search space

goal::Node{N}: goal position in the search space

low::SVector{N,Float64}: lower bounds of the search space

high::SVector{N,Float64}: upper bounds of the search space

goal_sample_rate::Float64: rate of sampling the goal node

step_size::Float64: maximum distance between each node

max_iter::Int64: maximum the number of the iterations

is_approved::Union{Function, Nothing}: a function that returns if the node can be added the graph

enable_logging::Bool: whether to log (is_goaled, RRTStar instance) for each steps in planning

PathPlanning.Planners.RRTStar — Type

RRTStar{N}

A planner using RRT*.

Fields

start::Node{N}: start position in the search space

goal::Node{N}: goal position in the search space

low::SVector{N,Float64}: lower bounds of the search space

high::SVector{N,Float64}: upper bounds of the search space

nodes::Vector{Node{N}}: nodes configure tje graph

goal_sample_rate::Float64: rate of sampling the goal node

step_size::Float64: maximum distance between each node

max_iter::Int64: maximum the number of the iterations

is_approved::Union{Function, Nothing}: a function that returns if the node can be added the graph

enable_logging::Bool: whether to log (is_goaled, RRTStar instance) for each steps in planning

logs::Vector{Dict{String, Any}}: (is_goaled, RRTStar instance) for each steps in planning

PathPlanning.Planners.RRTStar — Method

RRTStar(
    start::SVector{N,Float64},
    goal::SVector{N,Float64},
    low::SVector{N,Float64},
    high::SVector{N,Float64};
    goal_sample_rate::Float64 = 0.2,
    step_size::Union{Float64,Nothing} = nothing,
    max_iter::Int64 = 500,
    is_approved::Union{Function, Nothing} = nothing,
    enable_logging::Bool = false
) where {N}

Constructor of RRTStar struct. The step_size is initiated as 1/20th of the distance between the start node and the start node by default.

Arguments

start::Node{N}: start position in the search space

goal::Node{N}: goal position in the search space

low::SVector{N,Float64}: lower bounds of the search space

high::SVector{N,Float64}: upper bounds of the search space

goal_sample_rate::Float64: rate of sampling the goal node

step_size::Float64: maximum distance between each node

max_iter::Int64: maximum the number of the iterations

is_approved::Union{Function, Nothing}: a function that returns if the node can be added the graph

enable_logging::Bool: whether to log (is_goaled, RRTStar instance) for each steps in planning