Plot

PathPlanning.Plot.animate — Method

animate(
    env::Env,
    planner::AbstractPlanner{2};
    resolution::Tuple{Int64,Int64}=(1000,600),
    framerate::Int64=30,
    file_name::String="path.mp4"
)::Nothing

Create a movie that shows the planning process.

PathPlanning.Plot.plot — Method

plot(
    env::Env,
    planner::AbstractPlanner{2};
    resolution::Tuple{Int64,Int64}=(1000,600),
)::Tuple{Figure, Axis}

Plot nodes, paths, and obstacles that are in the env or the planner.

PathPlanning.Plot.plot_nodes! — Method

plot_nodes!(axis::Axis, nodes::Vector{Node{2}})::Nothing

Plot each nodes as small circles.

PathPlanning.Plot.plot_obstacle! — Method

plot_obstacle!(axis::Axis, obstacle::CircleObstacle{2})::Nothing

Plot a CircleObstacle.

PathPlanning.Plot.plot_obstacle! — Method

plot_obstacle!(axis::Axis, obstacle::RectObstacle{2})::Nothing

Plot a RectObstacle.

PathPlanning.Plot.plot_obstacles! — Method

plot_obstacles!(axis::Axis, obstacles::Vector{AbstractObstacle})::Nothing

Plot obstacles that can include RectObstacles, CircleObstacles, or both.

PathPlanning.Plot.plot_paths! — Method

plot_paths!(axis::Axis, nodes::Vector{Node{2}})::Nothing

Plot all paths as line segments. If you input all nodes that the planner has, you can plot all paths that the planner explorated, If you input nodes in a path that the planner found, you can plot a path that the planner found.